+++ R/align.R
@@ -9,43 +9,53 @@
 #' @param y2 The y-coordinate to align the second point to
 #' @param width The width of the aligned image
 #' @param height The height of the aligned image
-#' @param ref_img The reference image (by index or name) to get coordinates and dimensions from if NULL (defaults to first image)
+#' @param ref_img The reference image (by index or name) to get coordinates and dimensions from if they are NULL (defaults to average of all images if NULL)
 #' @param fill background color if cropping goes outside the original image
 #' @param patch whether to use the patch function to set the background color
-#' @param squash whether to move template points outside the image boundaries inside the image
 #' @param procrustes Whether to do a procrustes alignment
 #'
 #' @return stimlist with aligned tems and/or images
 #' @export
 #'
 #' @examples
-#' stimuli <- demo_stim("lisa")[1:2]
+#' # load stimuli and crop/rotate differently
+#' stimuli <- demo_stim() %>% 
+#'   crop(c(0.8, 0.9), c(1.0, 0.9), x_off = c(0, 0.2)) %>%
+#'   rotate(c(-5, + 5))
 #'
-#' # align to bottom-centre of nose
+#' # align to bottom-centre of nose (average position)
 #' one_pt <- align(stimuli, pt1 = 55, pt2 = 55)
 #'
 #' # align to pupils of second image
 #' two_pt <- align(stimuli, ref_img = 2)
 #'
-#' # procrustes align
+#' # procrustes align to average position
 #' proc <- align(stimuli, procrustes = TRUE)
+#' draw_tem(proc)
 
 align <- function(stimuli, pt1 = 0, pt2 = 1,
                   x1 = NULL, y1 = NULL, x2 = NULL, y2 = NULL,
-                  width = NULL, height = NULL, ref_img = 1,
-                  fill = wm_opts("fill"),
-                  patch = FALSE, squash = FALSE,
+                  width = NULL, height = NULL, ref_img = NULL,
+                  fill = wm_opts("fill"), patch = FALSE,
                   procrustes = FALSE) {
   stimuli <- validate_stimlist(stimuli, TRUE)
 
   # if values NULL, default to ref_img points
-  ref_points <- stimuli[[ref_img]]$points
+  if (is.null(ref_img)) {
+    avg <- average_tem(stimuli)
+    ref_points <- avg[[1]]$points
+    width <- width %||% width(avg)[[1]]
+    height <- height %||% height(avg)[[1]]
+  } else {
+    ref_points <- stimuli[[ref_img]]$points
+    width <- width %||% stimuli[[ref_img]]$width
+    height <- height %||% stimuli[[ref_img]]$height
+  }
   x1 <- x1 %||% ref_points[1, pt1+1]
   y1 <- y1 %||% ref_points[2, pt1+1]
   x2 <- x2 %||% ref_points[1, pt2+1]
   y2 <- y2 %||% ref_points[2, pt2+1]
-  width <- width %||% stimuli[[ref_img]]$width
-  height <- height %||% stimuli[[ref_img]]$height
+  
 
   if (pt1 == pt2) {
     # align points are the same, so no resize needed,
@@ -58,7 +68,7 @@ align <- function(stimuli, pt1 = 0, pt2 = 1,
   if (isTRUE(procrustes)) {
     # procrustes align coordinates
     data <- tems_to_array(stimuli)
-    coords <- procrustes_align(data)
+    coords <- procrustes_align(data, ref_img)
 
     # calculate size multiplier
     orig_avg <- apply(data, c(1, 2), mean)
@@ -68,7 +78,7 @@ align <- function(stimuli, pt1 = 0, pt2 = 1,
       sum() %>% sqrt(.)
     pEyeWidth <- (pro_avg[pt1+1, ] - pro_avg[pt2+1, ])^2 %>%
       sum() %>% sqrt(.)
-    mult = oEyeWidth/pEyeWidth
+    mult <- oEyeWidth/pEyeWidth
 
     # resize and convert to webmorph coordinates
     pts <- dim(coords)[[1]]
@@ -142,22 +152,40 @@ align <- function(stimuli, pt1 = 0, pt2 = 1,
   crop(newstimuli, width = width, height = height,
        x_off = x_off/width(newstimuli), # in case some values < 1
        y_off = y_off/height(newstimuli),
-       fill = fill, patch = patch, squash = squash)
+       fill = fill, patch = patch)
 }
 
 #' Procrustes align templates
 #'
 #' @param data Template points
-#' @param rotate Whether to "guess" the rotation, or rotate 0, 90, 180, or 270 degrees
+#' @param ref_img Reference image
 #'
 #' @return coordinates
 #' @keywords internal
 #'
-procrustes_align <- function(data, rotate = 0) {
-  gpa <- geomorph::gpagen(data, PrinAxes = FALSE,
-                          print.progress = FALSE)
+procrustes_align <- function(data, ref_img = NULL) {
+  n <- dim(data)[3]
+  if (is.null(ref_img)) {
+    # calcuate average and add as img 1
+    avg <- apply(data, c(1, 2), mean) %>% 
+      array(dim = dim(data) - c(0, 0, 1))
+    newdata <- array(c(avg, data), dim = dim(data) + c(0, 0, 1))
 
-  coords <- gpa$coords
+    data_i <- 2:(n+1)
+  } else {
+    # move ref image to first position
+    ref <- data[, , ref_img, drop = FALSE]
+    noref <- data[, , -ref_img, drop = FALSE]
+    newdata <- array(c(ref, noref), dim = dim(data))
+    if (ref_img == 1) {
+      data_i <- 1:n
+    } else {
+      data_i <- c(2:ref_img, 1, (ref_img+1):n)[1:n]
+    }
+  }
+  gpa <- geomorph::gpagen(newdata, PrinAxes = FALSE,
+                          print.progress = FALSE)
+  coords <- gpa$coords[, , data_i]
 
   # if (rotate != "guess") {
   #   rotate <- as.numeric(rotate)