+++ R/average_tem.R
@@ -1,6 +1,6 @@
 #' Average templates
 #'
-#' This function just averages the templates. An average image is returned, but it is just all the images superimposed. To create a full average, see \code{\link{avg}}.
+#' This function just averages the templates. An average image is returned, but it is just all the images superimposed. To create a template-aware average, see \code{\link{avg}}.
 #'
 #' @param stimuli list of class stimlist
 #' @param name Name for the average
@@ -9,7 +9,7 @@
 #' @export
 #'
 #' @examples
-#' demo_stim("lisa") %>% average_tem()
+#' demo_stim() %>% average_tem()
 #'
 average_tem <- function(stimuli, name = "average") {
   stimuli <- validate_stimlist(stimuli, TRUE)
@@ -19,13 +19,9 @@ average_tem <- function(stimuli, name = "average") {
 
   avg <- apply(pt, c(1, 2), mean)
 
-  for (i in seq_along(stimuli)) {
-    if (i == 1) {
-      img <- stimuli[[i]]$img
-    } else {
-      img <- c(img, stimuli[[i]]$img)
-    }
-  }
+  w <- width(stimuli) %>% mean()
+  h <- height(stimuli) %>% mean()
+  img <- crop(stimuli, w, h) %>% get_imgs()
 
   stim <- new_stim(
     magick::image_average(img),