+++ R/mirror.R
@@ -14,7 +14,7 @@
 #' m <- mirror(o, "frl")
 #'
 #' c(o, m) %>%
-#'   pt_subset(features("face")) %>%
+#'   subset_tem(features("face")) %>%
 #'   draw_tem(pt.shape = "index", pt.color = "white", pt.size = 10) %>%
 #'   label(c("original", "mirrored")) %>%
 #'   plot()